refactor: move private helper functions to helpers.go
